% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbin.R
\docType{package}
\name{rbin}
\alias{rbin}
\alias{rbin-package}
\title{\code{rbin} package}
\description{
Tools for binning data.
}
\details{
See the README on
\href{https://github.com/rsquaredacademy/rbin}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rsquaredacademy/rbin}
  \item \url{https://rbin.rsquaredacademy.com}
  \item Report bugs at \url{https://github.com/rsquaredacademy/rbin/issues}
}

}
\author{
\strong{Maintainer}: Aravind Hebbali \email{hebbali.aravind@gmail.com}

}
\keyword{internal}
