% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbin-create-bins.R
\name{rbin_create}
\alias{rbin_create}
\title{Create dummy variables}
\usage{
rbin_create(data, predictor, bins)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{predictor}{Variable for which dummy variables must be created.}

\item{bins}{An object of class \code{rbin_manual} or \code{rbin_quantiles} or \code{rbin_equal_length} or \code{rbin_winsorized}.}
}
\value{
\code{data} with dummy variables.
}
\description{
Create dummy variables from bins.
}
\examples{
\dontrun{
k <- rbin_manual(mbank, y, age, c(29, 31, 34, 36, 39, 42, 46, 51, 56))
rbin_create(mbank, age, k)
}

}
