% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbin-equal-freq.R
\name{rbin_equal_freq}
\alias{rbin_equal_freq}
\alias{plot.rbin_equal_freq}
\title{Equal frequency binning}
\usage{
rbin_equal_freq(data = NULL, response = NULL, predictor = NULL, bins = 10)

\method{plot}{rbin_equal_freq}(x, print_plot = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{response}{Response variable.}

\item{predictor}{Predictor variable.}

\item{bins}{Number of bins.}

\item{x}{An object of class \code{rbin_quantiles}.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\value{
A \code{tibble}.
}
\description{
Bin continuous data using the equal frequency binning method.
}
\examples{
bins <- rbin_equal_freq(mbank, y, age, 10)
bins

# plot
plot(bins)

}
