% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbin-factor.R
\name{rbin_factor}
\alias{rbin_factor}
\alias{plot.rbin_factor}
\title{Factor binning}
\usage{
rbin_factor(data = NULL, response = NULL, predictor = NULL,
  include_na = TRUE)

\method{plot}{rbin_factor}(x, print_plot = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{response}{Response variable.}

\item{predictor}{Predictor variable.}

\item{include_na}{logical; if \code{TRUE}, a separate bin is created for missing values.}

\item{x}{An object of class \code{rbin_factor}.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Weight of evidence and information value for categorical data.
}
\examples{
bins <- rbin_factor(mbank, y, education)
bins

# plot
plot(bins)

}
