% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbin-factor.R
\name{rbin_factor_combine}
\alias{rbin_factor_combine}
\title{Combine levels}
\usage{
rbin_factor_combine(data, var, new_var, new_name)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{var}{An object of class \code{factor}.}

\item{new_var}{A character vector; it should include the names of the levels to be combined.}

\item{new_name}{Name of the combined level.}
}
\value{
A \code{tibble}.
}
\description{
Manually combine levels of categorical data.
}
\examples{
upper <- c("secondary", "tertiary")
out <- rbin_factor_combine(mbank, education, upper, "upper")
table(out$education)

out <- rbin_factor_combine(mbank, education, c("secondary", "tertiary"), "upper")
table(out$education)

}
