% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbin-manual.R
\name{rbin_manual}
\alias{rbin_manual}
\title{Manual binning}
\usage{
rbin_manual(data = NULL, response = NULL, predictor = NULL,
  cut_points = NULL)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{response}{Response variable.}

\item{predictor}{Predictor variable.}

\item{cut_points}{Cut points for binning.}
}
\description{
Bin continuous variables manually.
}
\examples{
library(blorr)
rbin_manual(bank_marketing, y, age, c(29, 31, 34, 36, 39, 42, 46, 51, 56))

}
