% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbin-quantile.R
\name{rbin_quantiles}
\alias{rbin_quantiles}
\title{Quantile binning}
\usage{
rbin_quantiles(data = NULL, response = NULL, predictor = NULL,
  bins = 10)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{response}{Response variable.}

\item{predictor}{Predictor variable.}

\item{bins}{Number of bins}
}
\description{
Bin continuous variables using quantiles
}
\examples{
rbin_quantiles(marketing_bank, y, age, 10)


}
