% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbin-equal-length.R
\name{rbin_equal_length}
\alias{rbin_equal_length}
\alias{plot.rbin_equal_length}
\title{Equal length binning}
\usage{
rbin_equal_length(data = NULL, response = NULL, predictor = NULL,
  bins = 10, include_na = TRUE)

\method{plot}{rbin_equal_length}(x, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{response}{Response variable.}

\item{predictor}{Predictor variable.}

\item{bins}{Number of bins.}

\item{include_na}{logical; if \code{TRUE}, a separate bin is created for missing values.}

\item{x}{An object of class \code{rbin_equal_length}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A \code{tibble}.
}
\description{
Bin continuous data using the equal length binning method.
}
\examples{
\dontrun{
bins <- rbin_equal_length(mbank, y, age, 10)
bins

# plot
plot(bins)

}
}
