% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbin-equal-length.R
\name{rbin_equal_length}
\alias{rbin_equal_length}
\title{Equal length binning}
\usage{
rbin_equal_length(data = NULL, response = NULL, predictor = NULL,
  bins = 20)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{response}{Response variable.}

\item{predictor}{Predictor variable.}

\item{bins}{Number of bins.}
}
\description{
Bin continuous variables using the equal length binning method.
}
\examples{
library(blorr)
rbin_equal_length(bank_marketing, y, age, 20)

}
