% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbin-manual.R
\name{rbin_manual}
\alias{rbin_manual}
\alias{plot.rbin_manual}
\title{Manual binning}
\usage{
rbin_manual(data = NULL, response = NULL, predictor = NULL,
  cut_points = NULL)

\method{plot}{rbin_manual}(x, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{response}{Response variable.}

\item{predictor}{Predictor variable.}

\item{cut_points}{Cut points for binning.}

\item{x}{An object of class \code{rbin_manual}.}
}
\description{
Bin continuous variables manually.
}
\examples{
bins <- rbin_manual(marketing_bank, y, age, c(29, 31, 34, 36, 39, 42, 46, 51, 56))
bins

# plot
plot(bins)

}
