% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbin-addins.R
\name{rbinAddin}
\alias{rbinAddin}
\title{Bin variables}
\usage{
rbinAddin(data = NULL)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}
}
\description{
Manually bin variables using weight of evidence.
}
\examples{
\dontrun{
rbinAddin(data = mbank)
}

}
