% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbin-addins.R
\name{rbinFactorAddin}
\alias{rbinFactorAddin}
\title{Custom binning}
\usage{
rbinFactorAddin(data = NULL)
}
\description{
Manually combine categorical variables using weight of evidence.
}
\examples{
\dontrun{
rbinFactorAddin()
}

}
